Only specify encoding for sensitivity var if it exists
The sensitivity data var is optional, but `save_dataset_to_netcdf` fails
when it's absent because the encoding assumes it will be there, and
throws an exception when it finds it missing. This fixes that by only
adding the encoding for the sensitivity data var when it's present in
the dataset.