Reject zero-second duration audio segment intersections
The structure of the original conditional statement considered an audio
segment to be intersecting with a given minute if it ended the same
second the minute started. This results in an extra entry of zero-second
duration whenever the audio segment ends the same second one minute
ends and the next starts. Due to their zero-second duration, these
segments later cause a warning to be raised, but are ultimately ignored.
Rather than going through all that, this commit just restructures the
conditional statement to exclude entries whose intersecting duration with
the given minute is zero seconds.